[MTE-5838] - share the SwiftUI render helper between test files
Pulls the duplicated UIHostingController/UIWindow helper out of
GenericSelectableItemCellViewTests and NavigationBarMiddleButtonSelectionViewTests
into an XCTestCase extension so the two copies cannot drift.
