+++ firefox-ios/firefox-ios-tests/Tests/ClientTests/Settings/GenericSelectableItemCellViewTests.swift
@@ -49,16 +49,6 @@ final class GenericSelectableItemCellViewTests: XCTestCase {
         XCTAssertNotNil(render(subject))
     }
 
-    /// Lays the view out inside a window so that its body is evaluated.
-    private func render(_ view: some View) -> UIView {
-        let host = UIHostingController(rootView: view)
-        let window = UIWindow(frame: CGRect(x: 0, y: 0, width: 390, height: 844))
-        window.rootViewController = host
-        window.isHidden = false
-        host.view.layoutIfNeeded()
-        return host.view
-    }
-
     private func createSubject(isSelected: Bool,
                                theme: Theme? = LightTheme(),
                                onTap: @escaping () -> Void = {}) -> GenericSelectableItemCellView {
